+++ types/env.d.ts
@@ -1,6 +1,14 @@
-export interface Environment extends NodeJS.ProcessEnv {
+declare global {
+  namespace NodeJS {
+    interface ProcessEnv extends Environment {}
+  }
+}
+
+export interface Environment {
   NODE_ENV: "development" | "production";
   SECRET_KEY: string;
   PORT?: string | undefined;
+  DATABASE: string;
   DB_URI: string;
+  DB_URI_DEV: string;
 }